Add MSSQL db_port guard/unsupported, --verbose opt
Unfortunately, I couldn't find the correct api to use to make pyodbc
connect to an alternate port. As this is a new feature, and mysql tends
to be our primary audience, i'm justifying putting a DependencyError
guard there instead. We can revisit this in future.

--verbose flag was also added to log more things from the errors
produced my the main `pynonymizer` function. This should improve QoL for
debugging/troubleshooters.