Add fake `WHERE` to PG update to ensure randomness
Added an alias to all pg update statements, so we can refer to it in
subqueries that need to be run row-by-row, to "trick" the optimizer into
thinking that the generated value depends on the outer table.

TABLESAMPLE was not an option due to version constraints by users,
unfortunately, and it's worth pointing out that this STILL happens with
simple stuff like ( SELECT md5(random()::text) ) - that is, postgres
expects a subquery to be deterministc/cachable UNLESS it contains a
reference to something.