Remove database pre-test functionality
This might seem like a regression but upon closer inspection I'm convinced
this is the right course of action.

the pre-test functionality implies that one can test the connection
with the following conditions met:
1. you can connect to the server regardless of whether your target db is created:
  this is true in mssql and mysql, but not postgres.
2. connecting to the database is always possible, in all process types.
  There is some ongoing discussion on things like offline operation that would
  cast this into considerable doubt.

I don't consider removing it to be a breaking change because it was always an
invisible nice-to-have.