move logging to stderr, add "-" stdin/out
This is a bit of a change to the way logging is performed and it might
need some additional attention to make it seamless.

I'm treating stdin/out as files directly and using the filepath "-" to
denote the pipe. you should be able to use this like:

```
mysqldump | pynonymizer -i - -o - | other-process
```